Create AbstractPattern.js
An Abstract Factory creates objects that are related by a common theme. 
In object-oriented programming, a Factory is an object that creates other objects. An Abstract Factory has abstracted out a theme which is shared by the newly created objects.

JavaScript does not support class-based inheritance therefore the abstract classes as depicted in the diagram are not used in the JavaScript example. Abstract classes and interfaces enforce consistent interfaces in derived classes.